+++ Expanded Cheat Menu/Code/Menus/HelpFunc.lua
@@ -39,7 +39,7 @@ do -- ModUpload
 		ChoGGi_RocketsAutoLand = "Rockets Auto Land (upload 2)",
 		ChoGGi_SelectableCables = "Selectable Cables Pipes (upload 2)",
 		ChoGGi_StopColonistDeathFailure = "Stop Death Failure (upload 2)",
-		ChoGGi_UnitThoughts = "Unit Thoughts (upload 3)",
+		ChoGGi_UnitThoughts = "Unit Thoughts (upload 2)",
 		ChoGGi_WideAreaForestation = "Wide Area Forestation (upload 2)",
 	}
 	-- What? I like coloured titles...