+++ Mods ChoGGi/Remove Unwanted Colonists/metadata.lua
@@ -12,9 +12,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "1594867237",
 	"pops_any_uuid", "fb8fc954-40d2-4818-b84f-157903c4ed36",
 	"lua_revision", 1007000, -- Picard
-	"version", 20,
+	"version", 21,
 	"version_major", 2,
-	"version_minor", 0,
+	"version_minor", 1,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {