+++ Mods ChoGGi/Save Rocket Cargo/metadata.lua
@@ -1,10 +1,9 @@
 return PlaceObj("ModDef", {
 	"title", "Save Rocket Cargo",
-	"version", 4,
+	"version", 5,
 	"version_major", 0,
-	"version_minor", 4,
-
-	"image", "Preview.png",
+	"version_minor", 5,
+	"image", "Preview.jpg",
 	"id", "ChoGGi_SaveRocketCargo",
 	"steam_id", "1681523723",
 	"pops_any_uuid", "983cde10-49f0-4aab-b882-9061ca99b0ee",
@@ -14,7 +13,8 @@ return PlaceObj("ModDef", {
 		"Code/Script.lua",
 	},
 	"has_options", true,
-	"description", [[Stops rocket cargo from being removed when you leave the rocket screen.
+	"description", [[
+Stops rocket cargo from being removed when you leave the rocket screen.
 
 Closing the rocket dialog is when the cargo is saved, so switching from the pod cargo to the elevator will not save the pod cargo.
 The cargo will remain saved till you quit or load a different game.
@@ -23,5 +23,6 @@ There's a Clear button added to the cargo screen for "issues', see tooltip for m
 mod options:
 Clear On Launch: Clear cargo for rocket/pod/elevator when launched (not all cargo, just for the same type).
 
-Requested by... A bunch of people, but the last was Sigmatics.]],
+Requested by... A bunch of people, but the last was Sigmatics.
+]],
 })